# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 个人选菜清单 API
  version: 0.1.0
  description: |
    原生微信小程序和单管理员 Next.js 后台共用的版本化 API。
    所有金额均为人民币分的整数。私有资源的用户身份只能来自会话。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: 与管理后台或小程序配置域名相同的当前服务源
...
paths:
...
  /api/v1/admin/uploads/dish-image:
    post:
      tags: [Admin Dishes]
      summary: 上传待关联的菜品图片
      operationId: uploadDishImage
      security:
        - adminSession: []
          adminCsrf: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: JPEG、PNG 或 WebP，最大 5 MB
      responses:
        '201':
          description: 图片已安全保存，返回关联 ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-random-token
      description: wx.login 换取的随机令牌，最长有效 30 天
    adminSession:
      type: apiKey
      in: cookie
      name: ordering_admin_session
    adminCsrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  responses:
    ValidationError:
      description: 请求字段不符合契约
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthenticated:
      description: 会话缺失、无效或过期
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 已认证但禁止执行
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: 请求频率超过限制
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalError:
      description: 未预期服务端错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ResourceId:
      type: string
      minLength: 20
      maxLength: 32
      pattern: '^[a-z0-9]+$'
      examples: [cm1234567890abcdefghijk]
...
    ApiErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - WECHAT_UNAVAILABLE
        - INTERNAL_ERROR
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [ok, error]
      properties:
        ok: { const: false, type: boolean }
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { $ref: '#/components/schemas/ApiErrorCode' }
            message: { type: string, minLength: 1, maxLength: 200 }
            requestId: { type: string, minLength: 1, maxLength: 64 }
...
    Upload:
      type: object
      additionalProperties: false
      required: [id, mediaType, bytes, previewUrl]
      properties:
        id: { $ref: '#/components/schemas/ResourceId' }
        mediaType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        bytes: { type: integer, minimum: 1, maximum: 5242880 }
        previewUrl: { type: string, format: uri-reference }
    UploadResponse:
      type: object
      additionalProperties: false
      required: [ok, data]
      properties:
        ok: { const: true, type: boolean }
        data: { $ref: '#/components/schemas/Upload' }